feat: reorganize actions into separate files for bills and prescriptions
