+++ patient-management-system/src/app/(dashboard)/admin/fees/page.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import FeeForm from "@/app/(dashboard)/admin/fees/_components/FeeForm";
-import {getCharges} from "@/app/lib/actions";
+import {getCharges} from "@/app/lib/actions/charges";
 import {ChargeType} from "@prisma/client";
 
 const Page = async () => {
