+++ patient-management-system/src/app/(dashboard)/patients/[id]/prescriptions/[prescriptionID]/page.tsx
@@ -1,4 +1,5 @@
-import {getBill, getPrescription} from "@/app/lib/actions";
+import {getBill} from "@/app/lib/actions/bills";
+import {getPrescription} from "@/app/lib/actions/prescriptions";
 import {Card, CardContent, CardHeader} from "@/components/ui/card";
 import {CustomBadge} from "@/app/(dashboard)/_components/CustomBadge";
 import {Activity, Heart, HeartPulse} from "lucide-react";
