+++ patient-management-system/src/app/lib/actions.ts
@@ -2,10 +2,8 @@
 
 import {revalidatePath} from "next/cache";
 import {
-    Bill,
     DateRange,
     InventoryFormData,
-    myBillError,
     myError,
     PatientFormData,
     SortOption,
@@ -16,17 +14,11 @@ import {
 import {prisma} from "./prisma";
 import {verifySession} from "./sessions";
 import bcrypt from "bcryptjs";
-import {$Enums, BatchStatus, DrugType, Prisma, Role} from "@prisma/client";
-import {PrescriptionFormData} from "@/app/(dashboard)/patients/[id]/prescriptions/add/_components/PrescriptionForm";
-import {BrandOption} from "@/app/(dashboard)/patients/[id]/prescriptions/add/_components/IssueFromInventory";
-import {
-    BatchAssignPayload
-} from "@/app/(dashboard)/patients/[id]/prescriptions/[prescriptionID]/_components/BatchAssign";
+import { BatchStatus, DrugType, Prisma, Role} from "@prisma/client";
 import {ChangePasswordFormData} from "@/app/(dashboard)/admin/_components/ChangePasswordDialog";
 import {EditUserProfileFormData} from "@/app/(dashboard)/admin/_components/EditProfileDialog";
 import {validateEmail, validateMobile} from "@/app/lib/utils";
 import {AddUserFormData} from "@/app/(dashboard)/admin/staff/_components/AddUserDialog";
-import ChargeType = $Enums.ChargeType;
 import {SearchType} from "@/app/(dashboard)/queue/[id]/_components/CustomSearchSelect";
 
 export async function changePassword({
@@ -2453,796 +2445,5 @@ export async function getDrugModelStats(drugId: number) {
     } catch (error) {
         console.error("Error fetching stock analysis:", error);
         throw new Error("Failed to fetch stock analysis");
-    }
-}
-
-export async function addPrescription({
-                                          prescriptionForm,
-                                          patientID,
-                                      }: {
-    prescriptionForm: PrescriptionFormData;
-    patientID: number;
-}): Promise<myError> {
-    try {
-        // Basic validation checks
-        if (
-            prescriptionForm.issues.length === 0 &&
-            prescriptionForm.offRecordMeds.length === 0
-        ) {
-            return {
-                success: false,
-                message: "At least one prescription is required",
-            };
-        }
-
-        // check for repeated drugs
-        const drugIds = prescriptionForm.issues.map((issue) => issue.drugId);
-        const uniqueDrugIds = new Set(drugIds);
-        if (drugIds.length !== uniqueDrugIds.size) {
-            return {
-                success: false,
-                message:
-                    "Repeated drugs are not allowed in a single prescription",
-            };
-        }
-
-        // Create prescription with all related records in a transaction
-        await prisma.$transaction(async (tx) => {
-            // Create the main prescription and store its result to get issue IDs
-            const prescription = await tx.prescription.create({
-                data: {
-                    patientId: patientID,
-                    presentingSymptoms: prescriptionForm.presentingSymptoms,
-                    bloodPressure: prescriptionForm.bloodPressure,
-                    pulse: prescriptionForm.pulse,
-                    cardiovascular: prescriptionForm.cardiovascular,
-                    details: prescriptionForm.description,
-                    status: "PENDING",
-                    extraDoctorCharge: Number(
-                        prescriptionForm.extraDoctorCharges
-                    ),
-                    // Create issues
-                    issues: {
-                        create: prescriptionForm.issues.map((issue) => ({
-                            drugId: issue.drugId,
-                            details: issue.details,
-                            brandId: issue.brandId,
-                            strategy: issue.strategy,
-                            strategyDetails: issue.strategyDetails,
-                            quantity: issue.quantity,
-                        })),
-                    },
-                    // Create off-record medications
-                    OffRecordMeds: {
-                        create: prescriptionForm.offRecordMeds.map((med) => ({
-                            name: med.name,
-                            description: med.description,
-                        })),
-                    },
-                },
-                // Include the created issues in the return value
-                include: {
-                    issues: true,
-                },
-            });
-
-            //Check for queue entry
-            console.log(patientID);
-            const queueEntry = await tx.queueEntry.findFirst({
-                where: {
-                    patientId: patientID,
-                    status: "PENDING",
-                },
-            });
-
-            console.log(queueEntry);
-
-            if (queueEntry) {
-                revalidatePath(`/queue/${queueEntry.id}`);
-                await tx.queueEntry.update({
-                    where: {
-                        id: queueEntry.id,
-                    },
-                    data: {
-                        status: "PRESCRIBED",
-                    },
-                });
-            }
-
-            // Update strategy history for each issue
-            await Promise.all(
-                prescriptionForm.issues.map((issue, index) => {
-                    const createdIssue = prescription.issues[index]; // Get the corresponding created issue
-
-                    return tx.stratergyHistory.upsert({
-                        where: {
-                            drugId: issue.drugId,
-                        },
-                        update: {
-                            brandId: issue.brandId,
-                            issueId: createdIssue.id,
-                        },
-                        create: {
-                            drugId: issue.drugId,
-                            brandId: issue.brandId,
-                            issueId: createdIssue.id,
-                        },
-                    });
-                })
-            );
-        });
-
-        revalidatePath(`/patients/${patientID}/prescriptions`);
-        return {
-            success: true,
-            message: "Prescription created successfully",
-        };
-    } catch (e) {
-        console.error("Error adding prescription:", e);
-        return {
-            success: false,
-            message:
-                e instanceof Error
-                    ? e.message
-                    : "An error occurred while adding prescription",
-        };
-    }
-}
-
-export async function getCachedStrategy(drugID: number) {
-    return prisma.stratergyHistory.findUnique({
-        where: {
-            drugId: drugID,
-        },
-        select: {
-            issueId: true,
-            brandId: true,
-            issue: true,
-            brand: {
-                select: {
-                    name: true,
-                },
-            },
-        },
-    });
-}
-
-export async function searchAvailableDrugs(term: string) {
-    return prisma.drug
-        .findMany({
-            where: {
-                name: {
-                    startsWith: term,
-                },
-                batch: {
-                    some: {
-                        status: "AVAILABLE",
-                    },
-                },
-            },
-            take: 10,
-            select: {
-                id: true,
-                name: true,
-                _count: {
-                    select: {
-                        batch: true, // Counts the number of related batches
-                    },
-                },
-                batch: {
-                    distinct: ["drugBrandId"], // Get unique brands from batches
-                    select: {
-                        drugBrandId: true,
-                    },
-                },
-            },
-        })
-        .then((drugs) =>
-            drugs.map((drug) => ({
-                id: drug.id,
-                name: drug.name,
-                brandCount: drug.batch.length, // Count unique brands
-            }))
-        );
-}
-
-export async function searchBrandByDrug({
-                                            drugID,
-                                        }: {
-    drugID: number;
-}): Promise<BrandOption[]> {
-    return prisma.drugBrand
-        .findMany({
-            where: {
-                Batch: {
-                    some: {
-                        status: "AVAILABLE",
-                        drugId: drugID,
-                    },
-                },
-            },
-            select: {
-                id: true,
-                name: true,
-                Batch: {
-                    where: {
-                        status: "AVAILABLE",
-                        drugId: drugID,
-                    },
-                    select: {
-                        remainingQuantity: true,
-                        expiry: true,
-                        type: true,
-                    },
-                },
-            },
-        })
-        .then((brands) =>
-            brands.map((brand) => {
-                const batchCount = brand.Batch.length;
-                const totalRemainingQuantity = brand.Batch.reduce(
-                    (sum, batch) => sum + batch.remainingQuantity,
-                    0
-                );
-                const farthestExpiry = brand.Batch.reduce(
-                    (maxDate, batch) =>
-                        batch.expiry > maxDate ? batch.expiry : maxDate,
-                    new Date(0) // Initialize with the oldest possible date
-                );
-                return {
-                    id: brand.id,
-                    name: brand.name,
-                    batchCount,
-                    totalRemainingQuantity,
-                    farthestExpiry,
-                };
-            })
-        );
-}
-
-//For prescriptions page
-export async function getPrescriptionsCount(patientID: number) {
-    return prisma.prescription.count({
-        where: {
-            patientId: patientID,
-        },
-    });
-}
-
-export async function searchPrescriptions({
-                                              patientID,
-                                              query,
-                                              filter,
-                                              take,
-                                              skip,
-                                          }: {
-    patientID: number;
-    query: string;
-    filter: string;
-    take: number;
-    skip: number;
-}) {
-    let where = {};
-    const session = await verifySession();
-
-    if (query) {
-        if (filter === "symptom") {
-            where = {
-                presentingSymptoms: {
-                    contains: query,
-                },
-            };
-        }
-        if (filter === "drug") {
-            where = {
-                OR: [
-                    {
-                        issues: {
-                            some: {
-                                drug: {
-                                    name: {
-                                        contains: query,
-                                    },
-                                },
-                            },
-                        },
-                    },
-                    {
-                        OffRecordMeds: {
-                            some: {
-                                name: {
-                                    contains: query,
-                                },
-                            },
-                        },
-                    },
-                ],
-            };
-        }
-    }
-
-    // Only pending prescriptions can be viewed by non-doctor users
-    if (session.role !== "DOCTOR") {
-        where = {
-            ...where,
-            status: "PENDING",
-        };
-    }
-
-    return prisma.prescription.findMany({
-        where: {
-            patientId: patientID,
-            ...where,
-        },
-        take,
-        skip,
-        orderBy: {
-            time: "desc",
-        },
-        include: {
-            issues: {
-                include: {
-                    drug: true,
-                },
-            },
-            OffRecordMeds: {
-                select: {
-                    name: true,
-                },
-            },
-        },
-    });
-}
-
-export async function searchPrescriptionCount({
-                                                  patientID,
-                                                  query,
-                                                  filter,
-                                              }: {
-    patientID: number;
-    query: string;
-    filter: string;
-}) {
-    let where = {};
-
-    if (query) {
-        if (filter === "symptom") {
-            where = {
-                presentingSymptoms: {
-                    contains: query,
-                },
-            };
-        }
-        if (filter === "drug") {
-            where = {
-                OR: [
-                    {
-                        issues: {
-                            some: {
-                                drug: {
-                                    name: {
-                                        contains: query,
-                                    },
-                                },
-                            },
-                        },
-                    },
-                    {
-                        OffRecordMeds: {
-                            some: {
-                                name: {
-                                    contains: query,
-                                },
-                            },
-                        },
-                    },
-                ],
-            };
-        }
-    }
-
-    return prisma.prescription.count({
-        where: {
-            patientId: patientID,
-            ...where,
-        },
-    });
-}
-
-export async function getPrescription(
-    prescriptionID: number,
-    patientID: number
-) {
-    const session = await verifySession();
-    return prisma.prescription.findUnique({
-        where: {
-            id: prescriptionID,
-            patientId: patientID,
-            ...(session.role !== "DOCTOR" && {status: "PENDING"}),
-        },
-        include: {
-            issues: {
-                include: {
-                    drug: true,
-                    brand: true,
-                    batch: true,
-                },
-            },
-            OffRecordMeds: true,
-        },
-    });
-}
-
-export async function getBatches({
-                                     drugID,
-                                     brandID,
-                                 }: {
-    drugID: number;
-    brandID: number;
-}) {
-    return prisma.batch.findMany({
-        where: {
-            drugId: drugID,
-            drugBrandId: brandID,
-            status: "AVAILABLE",
-        },
-        select: {
-            id: true,
-            number: true,
-            remainingQuantity: true,
-            expiry: true,
-        },
-    });
-}
-
-export async function getCachedBatch({
-                                         drugId,
-                                         brandId,
-                                     }: {
-    drugId: number;
-    brandId: number;
-}) {
-    return prisma.batchHistory.findUnique({
-        where: {
-            drugId_drugBrandId: {
-                drugId,
-                drugBrandId: brandId,
-            },
-        },
-        select: {
-            batchId: true,
-        },
-    });
-}
-
-export async function calculateBill({
-                                        prescriptionData,
-                                    }: {
-    prescriptionData: BatchAssignPayload;
-}): Promise<myBillError> {
-    const prescriptionID = prescriptionData.prescriptionID;
-    const batchAssignments = prescriptionData.batchAssigns;
-
-    try {
-        return await prisma.$transaction(
-            async (prisma): Promise<myBillError> => {
-                const prescription = await prisma.prescription.findUnique({
-                    where: {id: prescriptionID},
-                });
-
-                if (!prescription) {
-                    return {
-                        success: false,
-                        message: "Prescription not found",
-                    };
-                }
-
-                if (prescription.status === "COMPLETED") {
-                    return {
-                        success: false,
-                        message: "Prescription already completed",
-                    };
-                }
-
-                let dspFees: number = 0;
-                const dispensaryFee = await prisma.charge.findUnique({
-                    where: {
-                        name: ChargeType.DISPENSARY,
-                    },
-                });
-
-                if (dispensaryFee) {
-                    dspFees = dispensaryFee.value;
-                }
-
-                let dctFee: number = 0;
-                const doctorFee = await prisma.charge.findUnique({
-                    where: {
-                        name: ChargeType.DOCTOR,
-                    },
-                });
-
-                if (doctorFee) {
-                    dctFee = doctorFee.value;
-                }
-
-                const bill: Bill = {
-                    patientID: prescriptionData.patientID,
-                    dispensary_charge: dspFees,
-                    doctor_charge:
-                        dctFee + (prescription.extraDoctorCharge ?? 0),
-                    cost: 0,
-                    entries: [],
-                };
-
-                for (let i = 0; i < batchAssignments.length; i++) {
-                    const assign = batchAssignments[i];
-                    if (!assign.batchID) {
-                        return {
-                            success: false,
-                            message: `Batch not found for a drug`,
-                        };
-                    }
-
-                    const batch = await prisma.batch.findUnique({
-                        where: {id: assign.batchID},
-                        include: {drug: true, drugBrand: true},
-                    });
-
-                    if (!batch) {
-                        return {
-                            success: false,
-                            message: `Batch not found for drug ${assign.batchID}`,
-                        };
-                    }
-
-                    const issue = await prisma.issue.findUnique({
-                        where: {id: assign.issueID},
-                    });
-
-                    if (!issue) {
-                        return {
-                            success: false,
-                            message: `Issue not found for drug ${assign.issueID}`,
-                        };
-                    }
-
-                    // Updating or creating the cache
-                    await prisma.batchHistory.upsert({
-                        where: {
-                            drugId_drugBrandId: {
-                                drugId: batch.drugId,
-                                drugBrandId: batch.drugBrandId,
-                            },
-                        },
-                        update: {
-                            batchId: assign.batchID,
-                        },
-                        create: {
-                            drugId: batch.drugId,
-                            drugBrandId: batch.drugBrandId,
-                            batchId: assign.batchID,
-                        },
-                    });
-
-                    await prisma.issue.update({
-                        where: {id: assign.issueID},
-                        data: {batchId: assign.batchID},
-                    });
-
-                    const batchCost = issue.quantity * batch.retailPrice;
-                    bill.cost += batchCost;
-
-                    bill.entries.push({
-                        drugName: batch.drug.name,
-                        brandName: batch.drugBrand.name,
-                        quantity: issue.quantity,
-                        unitPrice: batch.retailPrice,
-                    });
-                }
-
-                await prisma.bill.upsert({
-                    where: {
-                        prescriptionId: prescriptionID,
-                    },
-                    update: {
-                        doctorCharge: bill.doctor_charge,
-                        dispensaryCharge: bill.dispensary_charge,
-                        medicinesCharge: bill.cost,
-                    },
-                    create: {
-                        prescriptionId: prescriptionID,
-                        doctorCharge: bill.doctor_charge,
-                        dispensaryCharge: bill.dispensary_charge,
-                        medicinesCharge: bill.cost,
-                    },
-                });
-
-                return {
-                    success: true,
-                    message: "Bill calculated successfully",
-                    bill,
-                };
-            }
-        );
-    } catch (error) {
-        if (error instanceof Prisma.PrismaClientKnownRequestError) {
-            console.error(error.message);
-        }
-        return {
-            success: false,
-            message: "An error occurred while calculating bill",
-        };
-    }
-}
-
-export async function getBill(prescriptionID: number): Promise<Bill> {
-    const prescription = await prisma.prescription.findUnique({
-        where: {id: prescriptionID},
-        include: {
-            issues: {
-                include: {
-                    batch: true,
-                    drug: true,
-                    brand: true,
-                },
-            },
-            Bill: true,
-        },
-    });
-
-    if (!prescription) {
-        throw new Error("Prescription not found");
-    }
-
-    if (
-        prescription.status === "PENDING" ||
-        !prescription.Bill ||
-        !prescription.issues.every((issue) => issue.batch)
-    ) {
-        throw new Error("Prescription not completed");
-    }
-
-    return {
-        patientID: prescription.patientId,
-        dispensary_charge: prescription.Bill.dispensaryCharge,
-        doctor_charge: prescription.Bill.doctorCharge,
-        cost:
-            prescription.Bill.medicinesCharge +
-            prescription.Bill.dispensaryCharge +
-            prescription.Bill.doctorCharge,
-        entries: prescription.issues.map((issue) => ({
-            drugName: issue.drug.name,
-            brandName: issue.brand.name,
-            quantity: issue.quantity,
-            unitPrice: issue.batch?.retailPrice ?? 0,
-        })),
-    };
-}
-
-export async function completePrescription(
-    prescriptionID: number
-): Promise<myError> {
-    try {
-        const prescription = await prisma.prescription.findUnique({
-            where: {id: prescriptionID},
-            include: {
-                issues: {
-                    include: {
-                        batch: true,
-                    },
-                },
-            },
-        });
-
-        if (!prescription) {
-            return {success: false, message: "Prescription not found"};
-        }
-
-        if (prescription.status === "COMPLETED") {
-            return {
-                success: false,
-                message: "Prescription already completed",
-            };
-        }
-
-        if (!prescription.issues.every((issue) => issue.batch)) {
-            return {success: false, message: "Prescription not completed"};
-        }
-
-        // Updating remaining quantity of batches
-        await prisma.$transaction(async (prisma) => {
-            for (let i = 0; i < prescription.issues.length; i++) {
-                const issue = prescription.issues[i];
-                if (!issue.batch || !issue.batchId) {
-                    return {
-                        success: false,
-                        message: "Batch not found for a drug",
-                    };
-                }
-                await prisma.batch.update({
-                    where: {id: issue.batchId},
-                    data: {
-                        remainingQuantity: {
-                            decrement: issue.quantity,
-                        },
-                    },
-                });
-            }
-
-            await prisma.prescription.update({
-                where: {id: prescriptionID},
-                data: {status: "COMPLETED"},
-            });
-
-            //Check for queue entry
-            const queueEntry = await prisma.queueEntry.findFirst({
-                where: {
-                    patientId: prescription.patientId,
-                    status: "PRESCRIBED",
-                },
-            });
-
-            if (queueEntry) {
-                await prisma.queueEntry.update({
-                    where: {
-                        id: queueEntry.id,
-                    },
-                    data: {
-                        status: "COMPLETED",
-                    },
-                });
-                revalidatePath(`/queue/${queueEntry.id}`);
-            }
-        });
-        revalidatePath(`/patients/${prescription.patientId}/prescriptions`);
-        return {
-            success: true,
-            message: "Prescription completed successfully",
-        };
-    } catch (error) {
-        console.error("Error completing prescription:", error);
-        return {
-            success: false,
-            message: "An error occurred while completing prescription",
-        };
-    }
-}
-
-export async function getCharges() {
-    return prisma.charge.findMany({
-        where: {
-            OR: [{name: ChargeType.DISPENSARY}, {name: ChargeType.DOCTOR}],
-        },
-    });
-}
-
-export async function updateCharges({
-                                        charge,
-                                        value,
-                                    }: {
-    charge: ChargeType;
-    value: number;
-}): Promise<myError> {
-    try {
-        await prisma.charge.upsert({
-            where: {name: charge},
-            update: {value},
-            create: {name: charge, value},
-        });
-
-        return {
-            success: true,
-            message: `Charge for ${charge} updated successfully`,
-        };
-    } catch (e) {
-        if (e instanceof Prisma.PrismaClientKnownRequestError) {
-            console.error(e);
-        }
-
-        return {
-            success: false,
-            message: "An error occurred while updating charge",
-        };
     }
 }
